# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Reputation Management System API
  version: 1.0.0
  description: |
    API для системы управления репутацией компании. 
    Система позволяет автоматизировать сбор, анализ и мониторинг отзывов 
    о компаниях из различных источников (Яндекс, Google, 2GIS).
  contact:
    name: Reputation Management System
    email: support@reputation-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api
    description: Локальный сервер разработки
  - url: https://api.reputation-system.com/api
    description: Production сервер
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT токен аутентификации. Получите токен через /auth/login, затем используйте его в формате: Bearer {token}"
  schemas:
...
    CreateSourceRequest:
      type: object
      required:
        - companyId
        - platform
        - name
      properties:
        companyId:
          type: integer
        platform:
          type: string
          enum: [YANDEX, GOOGLE, TWOGIS]
        name:
          type: string
          maxLength: 200
        url:
          type: string
          maxLength: 500
        apiKey:
          type: string
          maxLength: 255
    SourceResponse:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        platform:
          type: string
          enum: [YANDEX, GOOGLE, TWOGIS]
        name:
          type: string
        url:
          type: string
        isActive:
          type: boolean
        lastCollection:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
        sourceId:
          type: integer
        externalId:
          type: string
        author:
          type: string
        rating:
          type: number
          format: double
          minimum: 1.0
          maximum: 5.0
        text:
          type: string
        publishedAt:
          type: string
          format: date-time
        collectedAt:
          type: string
          format: date-time
        sentiment:
          type: string
          enum: [POSITIVE, NEUTRAL, NEGATIVE]
        createdAt:
          type: string
          format: date-time
    UpdateSentimentRequest:
      type: object
      required:
        - sentiment
      properties:
        sentiment:
          type: string
          enum: [POSITIVE, NEUTRAL, NEGATIVE]
...
    CreateResponseRequest:
      type: object
      required:
        - taskId
        - text
      properties:
        taskId:
          type: integer
        text:
          type: string
    ResponseResponse:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        reviewId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        text:
          type: string
        recordedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /sources:
    post:
      tags:
        - Управление источниками
      summary: Создать источник
      description: Создание нового источника отзывов
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceRequest'
      responses:
        '201':
          description: Источник успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceResponse'
    get:
      tags:
        - Управление источниками
      summary: Получить все источники
      description: Получение списка всех источников отзывов
      responses:
        '200':
          description: Список источников
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceResponse'
  /sources/{id}:
    get:
      tags:
        - Управление источниками
      summary: Получить источник по ID
      description: Получение информации об источнике по идентификатору
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об источнике
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceResponse'
        '404':
          description: Источник не найден
  /sources/company/{companyId}:
    get:
      tags:
        - Управление источниками
      summary: Получить источники компании
      description: Получение списка источников для конкретной компании
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список источников компании
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceResponse'
  /sources/{id}/collect:
    post:
      tags:
        - Управление источниками
      summary: Запустить сбор отзывов
      description: Ручной запуск сбора отзывов из источника
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Сбор отзывов запущен
        '404':
          description: Источник не найден
...
  /reviews/{id}:
    get:
      tags:
        - Работа с отзывами
      summary: Получить отзыв по ID
      description: Получение информации об отзыве по идентификатору
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об отзыве
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '404':
          description: Отзыв не найден
    put:
      tags:
        - Работа с отзывами
      summary: Обновить тональность отзыва
      description: Обновление тональности отзыва
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSentimentRequest'
      responses:
        '200':
          description: Тональность успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
    delete:
      tags:
        - Работа с отзывами
      summary: Удалить отзыв
      description: Удаление отзыва из системы
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Отзыв успешно удален
        '404':
          description: Отзыв не найден
  /reviews/company/{companyId}:
    get:
      tags:
        - Работа с отзывами
      summary: Получить отзывы компании
      description: Получение списка отзывов для конкретной компании
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список отзывов компании
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
...
  /responses:
    post:
      tags:
        - Работа с ответами
      summary: Создать ответ
      description: Создание ответа на отзыв
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResponseRequest'
      responses:
        '201':
          description: Ответ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResponse'
    get:
      tags:
        - Работа с ответами
      summary: Получить все ответы
      description: Получение списка всех ответов
      responses:
        '200':
          description: Список ответов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseResponse'
  /responses/{id}:
    get:
      tags:
        - Работа с ответами
      summary: Получить ответ по ID
      description: Получение информации об ответе по идентификатору
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об ответе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResponse'
        '404':
          description: Ответ не найден
    delete:
      tags:
        - Работа с ответами
      summary: Удалить ответ
      description: Удаление ответа из системы
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Ответ успешно удален
        '404':
          description: Ответ не найден
  /responses/{id}/approve:
    put:
      tags:
        - Работа с ответами
      summary: Одобрить ответ
      description: Одобрение ответа модератором
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ответ одобрен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResponse'
        '404':
          description: Ответ не найден
  /responses/{id}/publish:
    put:
      tags:
        - Работа с ответами
      summary: Опубликовать ответ
      description: Публикация ответа на платформе
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ответ опубликован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseResponse'
        '404':
          description: Ответ не найден
...